Force build architecture to core2 on x86-64
This is to avoid that GMP does too aggresive choices when compiled on
AMD, failing to work when is then used on Corei7 or other Intel CPUs.

This will hopefully fix fix the error:

```
checking for double-to-integer conversion bug... yes or failed to exec (exit status is 132)
configure: error: double-to-integer conversion is incorrect.
You need to use another compiler (or lower the optimization level).
```

and similar compilation errors which are seen at random during the builds of
MPFR / FastJet. The joys of cloud computing.