Do not use zeromq on linux
The problem is that FairRoot does not allow for system ZeroMQ, so while
we can support it on mac (because we use brew) we have no means to find
where it is to be found on linux.

Once FairRoot properly validates and allows for ZeroMQ to come from the
system, we can drop this.