Fix ROOT6 tag to v6-06-04
Some of the recipes, most notably ROOT6 and ZeroMQ, track the head
of the branch rather than a specific tag. This was done because we
thought it was a good idea to track latest development from those (non
production) packages. Now that more people are starting to use aliBuild
it's clearly not a good default.