Fix SHERPA without wget and LHAPDF variable (#1)
SHERPA does not need `wget` any longer. The recipe now depends on `curl` and if
`wget` is not found on the system, then a "fake" wget script that in turn uses
`curl` is used. The script is used by SHERPA's `configure` to procure Sqlite3.
Note that the "fake wget" path comes after system's path so that system's `wget`
has precedence if available.