Adding ZeroMQ as dependency for AliRoot build in defaults-o2 (#743)
This handles https://github.com/AliceO2Group/AliceO2/issues/369

Note: AliRoot is by default disabled in the defaults, but might be
included in the build by removing this. Then the correct dependencies
must be applied.

In the override section, all dependencies have to be specified, simply
adding to the already defined dependencies is not yet supported.