Avoid conflict with local LHAPDF installation (#800)
Turns out LHAPDF is particularly stubborn in trying to pick up system
installation, if available. This results in errors when we are trying to
install our own PDF sets. This commit forces installation of PDF pdf
sets to use our own version of lhapdf, lhapdf-config (which was not the
case before) and forces usage of our own `share` path.