Make ROOT build when AliEn-Runtime is not there
Right now I think it simply works by accident, since the required dependencies
on libxml2 and OpenSSL are very likely to be found on the system. This should
make sure that we build our own version in case that's not the case.

We are also separating XRootD as, when AliEn-Runtime will be replaced with
jalien, we will maintain packages built independently instead of aggregating
them into a metapackage, for improved clarity.