Fix long shebangs in Python executables (#1057)
Python installation and setuptools produce by default Python scripts with the
full `python` executable path in the shebang. In some cases, on our installation
directories, that shebang exceeds 128 chars in length (max lenght on most
Linuxes), resulting in a "bad interpreter" error.

Since in our case we always have `python` in our environment, we now replace
every generated shebang with the shorter `/usr/bin/env python`.

We have noticed that this hack is not needed for the Python-modules package.

This complements #907.