Check for defaults_upper usage (#1140)
We should not use %(defaults_upper)s as part of the normal version
string because that makes a package and all it's dependencies defaults
dependent, even if there is no real need for it. If one needs a default
dependent behavior should add it to the overrides inside a default.