Use a working commit for FairRoot
This was already done but reverted citing our convention. Actually the convention states quite clearly:

> Do not create extra branches unless you do need to patch the original sources.

(see https://github.com/alisw/alidist/#guidelines-for-handling-externals-sources). Since this is the case here, I think using the hash is the correct thing to do (and regardless, according to @sawenzel the current tag is not the correct one).