Do not set MANPATH.
If MANPATH is set to simply o2 dir it will hide regular paths used by
man, so once that module is loaded a simple `man ls` will fail for
instance, which is pretty annoying.

man is smart enough to look in nearby directories of the directories in
PATH, so `man o2` will work just fine.