Speedup compilation (#1231)
`DYLD_LIBRARY_PATH` lookups are extremely slow on recent macOS version. Since `DYLD_LIBRARY_PATH` is not really needed at build time because we do the correct rpath relocation on macOS, we drop it from the O2 build instructions. We do not do it globally, because things like generators might actually require it to perform a correct build as well.