Fix ROOT plugin loading order for JAliEn (#1386)
This commit reverses order of loading external and internal plugins.
First, the external plugins are loaded, then internal. Order of loading
the plugins defines also the lookup order. The plugin priorities are
accross group, not accross all plugins available.

NOTE: If both external and internal plugin have the same base and
have the same regex, then the external plugin will be removed from
the plugin manager! However, the internal plugin will not replace it,
it will be registered at the end of list as any other internal plugin.

NOTE2: This modification should not affect behavior of any existing
ROOT plugin, is relevant only for JAliEn-ROOT and requires no further
modifications in ROOT.