Prevent setting CPLUS_INCLUDE_PATH on mac (#1671)
Fixing alisw/alidist#1656: In case boost and
python3 are taken from homebrew a collision
in the boost headers might happen as brew
installs a more recent boost version. As python
support is in any case currently disabled on
macOS dropping propagating the include path
for the python headers is safe on this platform.