Checkout base commit's lint-recipes script (#2767)
This fixes a security issue with the PR checker. Now, the scripts/lint-recipes
script from the PR's base commit is used, instead of from the PR's HEAD.

We also checkout the merge commit, not the latest commit in the PR, so we can
check if the PR integrates with changes to alidist made after the PR was opened.