coconut: force build requires to alibuild-generate-module
aliBuild does not consider as a build require something which is marked as
runtime require by one of the packages in the chain. This is because it cannot
know if you will use only the toplevel package (coconut, in this case) or one
it's dependencies (nothing stops you from loading gRPC directly, once you have
built coconut). However this does not play well with the automatic generation
of modulefiles. We use the FULL_BUILD_REQUIRES to correct the behavior of
alibuild-generate-module.