Fix xjalienfs shebang replacement
If we're building in a deeply-nested subdirectory, the scripts' shebangs will be
over 127 bytes, and therefore pip will use a hack to embed the python command
line, like:

    #!/bin/sh
    '''exec' /some/long/path/to/our/python3 "$0" "$@"
    ' '''

We want to replace this to use the system python3, but as that's a short
shebang, we can just use a regular #! line. The new command detects the above
case, deletes the first 3 lines, and inserts a regular #!/usr/bin/env python3
shebang instead.

If pip generates scripts with a regular shebang like:

    #!/some/shortish/path/to/our/python3

...we just replace the shebang with one pointing to the system python3.