Automatically open a PR from rc/ branches (#2735)
* Automatically open a PR from rc/ branches

With https://github.com/alisw/ali-bot/pull/910, daily-tags.sh pushes automatic
changes to an rc/* branch first to avoid overwriting others' changes. This
action creates a PR from that branch.

* Pass GitHub access token

Also, throw an error if the PR fails. The rc/ branch should exist as we don't
run when it was deleted.

* Fix and simplify script a little