Openloops 2.1.1 - introduce python as requirement
On Kubuntu 20.04 LTS (i.e. non-slc7, non-macOS), Openloops does not compile due a failing verification of a SSL certificate while calling the python script: Openloops/pyol/bin/download_process.py.

2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: ++ Line 15: PROCESSES=(ppjj ppjj_ew ppjjj ppjjj_ew ppjjj_nf5 ppjjjj)
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: ++ Line 16: for proc in ${PROCESSES[@]}
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: ++ Line 17: ./scons --jobs=2 auto=ppjj
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: scons: Reading SConscript files ...
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: 
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: >>> OpenLoops Process Downloader <<<
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: 
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: Warning: Channel database update for repository public failed (<urlopen error [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed (_ssl.c:727)>). Skip this repository.
2022-08-08@16:41:11:DEBUG:AliGenerators:Openloops:vAN-2022: - process: ppjj ... ERROR: not available (installed: {})

The downloads are needed to get desired NLO physics processes for our needs (ex : ppjj = pp to dijets, etc), such a shopping list is defined in our recipe alidist/openloops.sh.

What to do ?

A fix for this can be in place in the recipe side by requesting an explicit built-time and run-time dependency (requires) to Python package not only for slc platforms, but rather anything which is non-MacOs (e.g. Ubuntu).
Note that a similar policy has been set up recently by Peter Hristov for yoda.sh in PR 4408.
The trick is apparently in the alidist/python.sh (l.91) where a SSL certificate are managed and anticipated properly.
Then requesting python as requirement for Openloops cures the thing on the way.