alidist/YODA - introduce (py)ROOT dependency
(py)ROOT can be introduced as dependency to YODA.
The feature is at least needed in order get a feature allowing the user to move from one format to the next between YODA and ROOT.
A corresponding (python+cython) package is to be built within YODA + see enduser scripts YODA/bin/root2yoda + YODA/bin/yoda2root

Mods in yoda.sh recipe:
1. The recipe now _requires_ ROOT (pyROOT as a matter of fact).
2. The _configure_ needs to enable the (py)ROOT compatibility : ./configure --disable-silent-rules --enable-root (for any platform : macOS, slc, Ubuntu...)
3. The YODA _module_ finally needs to load ROOT along.

- Extra -

The commit also content notes and 2 questions about Python management in the recipe.
This is for my own education (PAG-MC).
The recipe is used here as a playground; expert views are needed for clarification.
Comments put in the recipe will likely be removed in a next commit,
while information got here from exchange with experts may be moved to aliBuild documentation :
https://github.com/alisw/alibuild/tree/master/docs

A very similar python interrogation runs in parallel for the lhapdf.sh recipe, where python is also needed as recipe requirement.