Export ALICE_PHYSICS regardless of default (#4485)
ALICE_PHYSICS is currently exported by the aliphysics.sh. However, due to the fact defaults-o2 overrides aliphysics attributes without adding back ALICE_PHYSICS, the gun test fails. Given in the end such a variable is only needed by the guntest, the best solution is to simply export it explicitly there from ALIPHYSICS_ROOT, which is guaranteed to be always there.