Install pkgconfig file for UUID (#4645)
* Install pkgconfig file for UUID

XRootD requires this, as it searches for libuuid using pkg-config in its CMake
recipes. This leads to build failures on Ubuntu. While users can install libuuid
on their system to provide the `*.pc` file, this seems silly if we're building
the library anyway.

* Use the more common prepend_path

This makes sure our package is used ahead of system ones, if the path is already
defined by the user.