Removed edit committes on 23/09/2020
After the merge, from that day on it was not possible anymore to run RIVET on the GRID. I believe this is the only modification performed. 
So a test could be valuable to understand if this will fix the issue. 
As of now only the last version runnable on the GRID is the one of 22-09-2020