Fix mistakes in YAML headers
defaults-o2-dataflow.sh: ON and OFF are parsed as booleans in YAML, but should
be strings.

fairmq.sh: The second ROOT_INCLUDE_PATH overrides the first one, but they should
very likely both be included.

geant4.sh: Don't manually fiddle with colons in paths, but let aliBuild do it.

openmp.sh: 1.0 is parsed as a float in YAML, but should be a string.

parquet-cpp.sh: The second build_requires overrides the first one, but they
should be combined instead.

rivettask.sh: %{x}s is not a valid formatting specifier; use %(x)s instead.

template-recipe.sh: The package name should match the file name.