Re-enable simulation in O2 builds; find geant4 + xercesc (#5052)
Since the introduction of GDML support in Geant4,
some builds of O2 failed to include simulation. This lead
to failure of the fullCI and absence of o2-sim on the GRID.

The problem is that finding Geant4 during O2-cmake triggers again
a search for XercesC, which seems to fail.

This commit fixes the issue by asking cmake to prefer CONFIG package searches
which guides it better to pickup XercesC from the alibuild software stack.

This appears to be the shortest possible (hot)fix for the issue.