Revert removing llvm-spirv build from clang
Just found out by chance that the internal clang spirv backend is not working.
(Not at all in LLVM15, and in LLVM17 it still produces broken code.)
Thus, for now we have to keep bundling the external tool with clang.
So far, O2 was building with partial OpenCL compilation, due to a bug in
the O2 CMake, which silently disabled the SPIRV translation in this case.