Use of HepMC3 instead of HepMC2 (#5206)
* Update thepeg.sh

Use HepMC3 instead of HepMC (version 2).  For Rivet updates to use HepMC3 instead of old and deprecated HepMC (version 2).

* Update thepeg.sh

Asked to bump down tag - the patch level of the tag was increased because the previous tag didn't work due to problem with required HepMC version.   I thought it best to bump the patch-level so as to not create confusion.