Fixed issue with CGAL path substitution
I had

```
fastjet-config --libs | \
  tr ' ' '\n'  | \
  grep cgal | \
  sed -n 's|-L\(.*\)/cgal.*|\1|p'
```

which would do

```
/cvmfs/alice.cern.ch/el7-x86_64/Packages/cgal/4.6.3-131/lib =>
/cvmfs/alice.cern.ch/el7-x86_64/Packages =>
```

which would then leave the `cgal/4.6.3-131/lib` part in the
`rivet-config` and `rivet-build` scripts, which of course shouldn't be
there.  The right extraction should be

```
fastjet-config --libs | \
  tr ' ' '\n'  | \
  grep cgal | \
  sed -n 's|-L\(.*\)/lib|\1|p'
```

In this commit I also copy the original script (but disable the execute
bit) for debugging purposes.  We can of course remove that once we're
happy.