Create a new "grid-base-packages" modulefile
This will be loaded by the toolchain on CVMFS, so that we have a common set of
required packages for Grid use available. Currently, only jq is needed.

A separate PR will be needed to load this package from GCC-Toolchain's
modulefile.

A "default" symlink must also be created for the modulefiles of this package
on CVMFS.