Load grid-base-packages/default in toolchain module
For use on the Grid and e.g. in Hyperloop, some packages need to be available
(currently just jq). Package these as "grid-base-packages", which has a
"default" symlink available.

Then, we can add more packages in future that should be loaded as part of the
toolchain, without incurring costly rebuilds of the full software stack.