Fix libuv load path (#5429)
libwebsockets is linked statically, however it requires libuv, which is not, therefore one needs to actually have it in the modulefile. This works just fine on O2, because other stuff brings in libuv, however it breaks when using it for AliROOT, which does not. Hot patched on CVMFS as well.