STARlight introduction in AliGenerators
Added STARlight dependency to AliGenerators. STARlight recipe has been edited to pull directly the main branch from the official STARlight repository. In addition the HepMC3 package has been set as a dependency and STARlight is now built with it, so that it could also be useful for RIVET analyses. Tests performed locally are encouraging.