Correct handling of RPATH for grpc and c-ares (#5561)
For some reason c-ares and grpc set their id to `@rpath/lib...` rather
than their full path. This means that if a given dependent library has
an implicit transitive dependency on it, they will not be able to find it.

This happens in particular for bookkeeping which defines grpc as a private
dependency, breaking downstream users.