Lock ROOT and FairMQ versions for online environment
We don't want to change the versions of FairMQ and ROOT in the online environment until the end of the 2025 run. On the other hand, bumps of those packages are needed in other setups (e.g. async). 

This change ensures that when we build the RPMs for the FLPs we keep the current versions.