Forced use of pre-built fastjet and zlib
Fastjet requires CGAL and GMP to work, otherwise the build will fail. ROOT was an added dependency otherwise MadAnalysis5 would have complained that TH1F was missing. Lastly the trick with ZLIB was necessary because HepTools checks if zlib is available ONLY in the HepTools folder, and not if it is provided as a path. Since changing the last behaviour would require a much longer and complicated patch (HepTools is an external package downloaded with curl), this seemed to be the easiest option