Upgrade and include Herwig on AliGenO2 (#5997)
This integrates Herwig into AliGenO2 so that it will get published on CVMFS.

Fastjet requires CGAL and GMP to work, otherwise the build will fail. ROOT was an added dependency otherwise MadAnalysis5 would have complained that TH1F was missing. Lastly the trick with ZLIB was necessary because HepTools checks if zlib is available ONLY in the HepTools folder, and not if it is provided as a path. Since changing the last behaviour would require a much longer and complicated patch (HepTools is an external package downloaded with curl), this seemed to be the easiest option

* Include header folders in ROOT_INCLUDE_PATH
* ThePEG compatibility update

If ThePEG was used with O2DPG JIT compiled macros, the simulation was crashing. The new version fixes the issue, bypassing the HepMC namespace conversion by setting a preprocessor variable and providing the correct GenCrossSection function in the HepMCTraits